Change name class
