+++ goodprice/demo/src/test/java/com/example/test/core/ainhconfigs.properties
@@ -1,0 +1,3 @@
+browser=CHROME
+url=https://staging.kbb.com/cars-for-sale/all
+
