--- goodprice/demo/src/test/java/com/example/test/core/configs.properties
@@ -1,3 +1,0 @@
-browser=CHROME
-url=https://staging.kbb.com/cars-for-sale/all
-
